Adds user-asserted author data on export screen to teidom, api, and pdf
Defines a few constants ANTHOLOGIZE_CREATORS_ASSERTED and ANTHOLOGIZE_CREATORS_ALL as a start to managing the options for which authors can and should be displayed. Adds the data into the TEI, and implements the switch in pdf to choose between them. Default behavior is what we had before -- just the creator as the WP project creator.