Fixes #50
Mostly, at least. Layout is still different. Might be a function of
new stripping out of align and other attributes, or just a gap in what
TCPDF can work with. More detailed exploration needed there.

This also gets us toward being ready to get rid of the shortcodes
enable/disable option in all the output formats. That'll call for more
testing against the default WP shortcodes, but it's a good step toward
removing clutter that does not necessarily do what it purports to do.