More graceful failure if ePub system requirements are not met.
Previously, it was possible to run through the entire export process before
discovering that the ePub extension was not available (due to the lack of
XSL support). This message was delivered via a whitescreen, at the very last
stage of export.

Now, when Anthologize detects that XSL is not present on the system, two things
happen:

1. If the user is an administrator (`current_user_can( 'install_plugins' )`),
   the ePub option is visible but disabled, with a message explaining why it
   cannot be selected.
2. If the user is not an administrator, ePub will not appear as an option.