Use source item author names to generate default project author string.
The 'Author(s)' field on the Export Project panel is now prefilled
with a comma-separated list of the display names belonging to the authors
of the source content. This can then be modified by the exporter.

As part of this change, the formatting of author info in PDF output
has been improved, to ensure centering of long text.