Don't clone post attachments when creating a new library item.
Introduced in #50, this block is not necessary for several reasons. First,
by cloning the attachment post without also bringing over its postmeta -
specifically the attachment-specific metadata - it created media items that
WP was not able to render fully in the Media Library. Second, the cloned
attachments were, in a sense, orphaned, since numeric ID references in
gallery shortcodes and elsewhere were not swapped out for the IDs of the
cloned items. Third and most important, it's not necessary to clone post
attachments to make galleries etc work, because a library item can freely
make reference to media items that don't have the library item as
`post_parent` (either by post ID or by URL).

Fixes #113.